requirements.txt
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.0
Django==3.2.3
djangorestframework==3.12.4
flake8==7.1.1
idna==3.10
mccabe==0.7.0
pycodestyle==2.12.1
pyflakes==3.2.0
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
